requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

yfinance==0.2.28
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
pandas==2.1.4
numpy==1.26.2

openai==1.3.7
faiss-cpu==1.7.4
python-dotenv==1.0.0

openai-whisper==20231117
pyttsx3==2.90
SpeechRecognition==3.10.0
gTTS==2.4.0

pydantic==2.5.2
python-multipart==0.0.6

sentence-transformers==2.2.2
plotly==5.17.0
scikit-learn==1.3.2

